updated sqlhandler
