+++ SQLHelper/SQLHandlerAsync.cs
@@ -15,9 +15,9 @@ namespace SQLHelper
         public static string Connectionconfig;
         #region "Private Members"
 
-        private string _objectQualifier = string.Empty;
-        private string _databaseOwner = "[dbo]";
-        private string _connectionString = "";
+        internal string _objectQualifier = string.Empty;
+        internal string _databaseOwner = "[dbo]";
+        internal string _connectionString = "";
 
         #endregion
 
@@ -109,54 +109,8 @@ namespace SQLHelper
             return;
         }
 
-        public int ExecuteNonQuery(SqlTransaction transaction, CommandType commandType, string commandText, List<KeyValuePair<string, object>> ParaMeterCollection, string outParamName)
-        {
-            //create a command and prepare it for execution
-            SqlCommand cmd = new SqlCommand();
-            PrepareCommand(cmd, transaction.Connection, transaction, commandType, commandText);
 
-            for (int i = 0; i < ParaMeterCollection.Count; i++)
-            {
-                SqlParameter sqlParaMeter = new SqlParameter();
-                sqlParaMeter.IsNullable = true;
-                sqlParaMeter.ParameterName = ParaMeterCollection[i].Key;
-                sqlParaMeter.Value = ParaMeterCollection[i].Value;
-                cmd.Parameters.Add(sqlParaMeter);
-            }
-            cmd.Parameters.Add(new SqlParameter(outParamName, SqlDbType.Int));
-            cmd.Parameters[outParamName].Direction = ParameterDirection.Output;
 
-            //finally, execute the command.
-            cmd.ExecuteNonQuery();
-            int id = (int)cmd.Parameters[outParamName].Value;
-
-            // detach the OracleParameters from the command object, so they can be used again.
-            cmd.Parameters.Clear();
-            return id;
-        }
-
-        public void ExecuteNonQuery(SqlTransaction transaction, CommandType commandType, string commandText, List<KeyValuePair<string, object>> ParaMeterCollection)
-        {
-            //create a command and prepare it for execution
-            SqlCommand cmd = new SqlCommand();
-            PrepareCommand(cmd, transaction.Connection, transaction, commandType, commandText);
-
-            for (int i = 0; i < ParaMeterCollection.Count; i++)
-            {
-                SqlParameter sqlParaMeter = new SqlParameter();
-                sqlParaMeter.IsNullable = true;
-                sqlParaMeter.ParameterName = ParaMeterCollection[i].Key;
-                sqlParaMeter.Value = ParaMeterCollection[i].Value;
-                cmd.Parameters.Add(sqlParaMeter);
-            }
-
-            //finally, execute the command.
-            cmd.ExecuteNonQuery();
-
-            // detach the OracleParameters from the command object, so they can be used again.
-            cmd.Parameters.Clear();
-
-        }
         #endregion
 
         #endregion
@@ -389,313 +343,6 @@ namespace SQLHelper
             }
         }
         
-        /// <summary>
-        /// Returning Bool After Execute Non Query
-        /// </summary>
-        /// <param name="StroredProcedureName">Store Procedure Name</param>
-        /// <param name="ParaMeterCollection"> Parameter Collection</param>
-        /// <param name="OutPutParamerterName">Out Parameter Collection</param>
-        /// <returns>Bool</returns>
-        public async Task<bool> ExecuteNonQueryAsBoolAsync(string StroredProcedureName, List<KeyValuePair<string, object>> ParaMeterCollection, string OutPutParamerterName)
-        {
-            using (SqlConnection SQLConn = new SqlConnection(this._connectionString))
-            {
-                using (SqlCommand SQLCmd = new SqlCommand())
-                {
-                    SQLCmd.Connection = SQLConn;
-                    SQLCmd.CommandText = StroredProcedureName;
-                    SQLCmd.CommandType = CommandType.StoredProcedure;
-                    //Loop for Paramets
-                    for (int i = 0; i < ParaMeterCollection.Count; i++)
-                    {
-                        SqlParameter sqlParaMeter = new SqlParameter();
-                        sqlParaMeter.IsNullable = true;
-                        sqlParaMeter.ParameterName = ParaMeterCollection[i].Key;
-                        sqlParaMeter.Value = ParaMeterCollection[i].Value;
-                        SQLCmd.Parameters.Add(sqlParaMeter);
-                    }
-                    //End of for loop
-                    SQLCmd.Parameters.Add(new SqlParameter(OutPutParamerterName, SqlDbType.Bit));
-                    SQLCmd.Parameters[OutPutParamerterName].Direction = ParameterDirection.Output;
-                    try
-                    {
-                        await SQLConn.OpenAsync();
-                        await SQLCmd.ExecuteNonQueryAsync();
-                        bool ReturnValue = (bool)SQLCmd.Parameters[OutPutParamerterName].Value;
-
-                        return ReturnValue;
-                    }
-                    catch (Exception e)
-                    {
-                        throw e;
-                    }
-                    finally
-                    {
-                        SQLConn.Close();
-                    }
-                }
-
-            }
-        }
-
-        /// <summary>
-        /// Returning Bool After Execute Non Query
-        /// </summary>
-        /// <param name="StroredProcedureName">Store Procedure Name</param>
-        /// <param name="ParaMeterCollection">Parameter Collection</param>
-        /// <param name="OutPutParamerterName">OutPut Parameter Name</param>
-        /// <param name="OutPutParamerterValue">OutPut Parameter Value</param>
-        /// <returns>Bool</returns>
-        public async Task<bool> ExecuteNonQueryAsBoolAsync(string StroredProcedureName, List<KeyValuePair<string, object>> ParaMeterCollection, string OutPutParamerterName, object OutPutParamerterValue)
-        {
-            using (SqlConnection SQLConn = new SqlConnection(this._connectionString))
-            {
-
-                using (SqlCommand SQLCmd = new SqlCommand())
-                {
-                    SQLCmd.Connection = SQLConn;
-                    SQLCmd.CommandText = StroredProcedureName;
-                    SQLCmd.CommandType = CommandType.StoredProcedure;
-                    //Loop for Paramets
-                    for (int i = 0; i < ParaMeterCollection.Count; i++)
-                    {
-                        SqlParameter sqlParaMeter = new SqlParameter();
-                        sqlParaMeter.IsNullable = true;
-                        sqlParaMeter.ParameterName = ParaMeterCollection[i].Key;
-                        sqlParaMeter.Value = ParaMeterCollection[i].Value;
-                        SQLCmd.Parameters.Add(sqlParaMeter);
-                    }
-                    //End of for loop
-                    SQLCmd.Parameters.Add(new SqlParameter(OutPutParamerterName, SqlDbType.Bit));
-                    SQLCmd.Parameters[OutPutParamerterName].Direction = ParameterDirection.Output;
-                    SQLCmd.Parameters[OutPutParamerterName].Value = OutPutParamerterValue;
-                    try
-                    {
-                        await SQLConn.OpenAsync();
-                        await SQLCmd.ExecuteNonQueryAsync();
-                        bool ReturnValue = (bool)SQLCmd.Parameters[OutPutParamerterName].Value;
-                        return ReturnValue;
-                    }
-                    catch (Exception e)
-                    {
-                        throw e;
-                    }
-                    finally
-                    {
-                        SQLConn.Close();
-                    }
-                }
-
-            }
-        }
-
-        /// <summary>
-        /// Execute Non Query
-        /// </summary>
-        /// <param name="StroredProcedureName">Store Procedure Name In String</param>
-        /// <param name="ParaMeterCollection">Accept Key Value Collection For Parameters<KeyValuePair<string, object>> </param>
-        public async Task<int> ExecuteNonQueryAsync(string StroredProcedureName, List<KeyValuePair<string, object>> ParaMeterCollection)
-        {
-            using (SqlConnection SQLConn = new SqlConnection(this._connectionString))
-            {
-
-                using (SqlCommand SQLCmd = new SqlCommand())
-                {
-                    SQLCmd.Connection = SQLConn;
-                    SQLCmd.CommandText = StroredProcedureName;
-                    SQLCmd.CommandType = CommandType.StoredProcedure;
-                    //Loop for Paramets
-                    for (int i = 0; i < ParaMeterCollection.Count; i++)
-                    {
-                        SqlParameter sqlParaMeter = new SqlParameter();
-                        sqlParaMeter.IsNullable = true;
-                        sqlParaMeter.ParameterName = ParaMeterCollection[i].Key;
-                        sqlParaMeter.Value = ParaMeterCollection[i].Value;
-                        SQLCmd.Parameters.Add(sqlParaMeter);
-                    }
-                    //End of for loop
-                    try
-                    {
-                        int effectedRows = 0;
-                        await SQLConn.OpenAsync();
-                        effectedRows = await SQLCmd.ExecuteNonQueryAsync();
-                        return effectedRows;
-                    }
-                    catch (Exception e)
-                    {
-                        throw e;
-                    }
-                    finally
-                    {
-                        SQLConn.Close();
-                    }
-                }
-
-            }
-        }
-
-        /// <summary>
-        /// Execute Non Query
-        /// </summary>
-        /// <param name="StroredProcedureName">Store Procedure Name</param>
-        /// <param name="ParaMeterCollection">Accept Key Value Collection For Parameters <KeyValuePair<string, string>> </param>
-        public async Task<int> ExecuteNonQueryAsync(string StroredProcedureName, List<KeyValuePair<string, string>> ParaMeterCollection)
-        {
-            using (SqlConnection SQLConn = new SqlConnection(this._connectionString))
-            {
-
-                using (SqlCommand SQLCmd = new SqlCommand())
-                {
-                    SQLCmd.Connection = SQLConn;
-                    SQLCmd.CommandText = StroredProcedureName;
-                    SQLCmd.CommandType = CommandType.StoredProcedure;
-                    //Loop for Paramets
-                    for (int i = 0; i < ParaMeterCollection.Count; i++)
-                    {
-                        SQLCmd.Parameters.Add(new SqlParameter(ParaMeterCollection[i].Key, ParaMeterCollection[i].Value));
-                    }
-                    //End of for loop
-                    try
-                    {
-                        int effectedRows = 0;
-                        await SQLConn.OpenAsync();
-                        effectedRows = await SQLCmd.ExecuteNonQueryAsync();
-                        return effectedRows;
-                    }
-                    catch (Exception e)
-                    {
-                        throw e;
-                    }
-                    finally
-                    {
-                        SQLConn.Close();
-                    }
-                }
-            }
-
-        }
-
-        /// <summary>
-        /// Execute Non Query
-        /// </summary>
-        /// <param name="StroredProcedureName">Store Procedure Name</param>
-        public async Task<int> ExecuteNonQueryAsync(string StroredProcedureName)
-        {
-            using (SqlConnection SQLConn = new SqlConnection(this._connectionString))
-            {
-                using (SqlCommand SQLCmd = new SqlCommand())
-                {
-                    SQLCmd.Connection = SQLConn;
-                    SQLCmd.CommandText = StroredProcedureName;
-                    SQLCmd.CommandType = CommandType.StoredProcedure;
-                    try
-                    {
-                        int effectedRows = 0;
-                        await SQLConn.OpenAsync();
-                        effectedRows = await SQLCmd.ExecuteNonQueryAsync();
-                        return effectedRows;
-                    }
-                    catch (Exception e)
-                    {
-                        throw e;
-                    }
-                    finally
-                    {
-                        SQLConn.Close();
-                    }
-                }
-            }
-        }
-
-        /// <summary>
-        /// Accept only int, Int16, long, DateTime, string (NVarcha of size  50),
-        /// bool, decimal ( of size 16,2), float
-        /// </summary>
-        /// <typeparam name="T">Return the given type of object</typeparam>
-        /// <param name="StroredProcedureName">Accet SQL procedure name in string</param>
-        /// <param name="ParaMeterCollection">Accept Key Value Collection for parameters</param>
-        /// <param name="OutPutParamerterName">Accept Output parameter for the stored procedures</param>
-        /// <returns></returns>
-        public async Task<T> ExecuteNonQueryAsGivenTypeAsync<T>(string StroredProcedureName, List<KeyValuePair<string, object>> ParaMeterCollection, string OutPutParamerterName)
-        {
-            using (SqlConnection SQLConn = new SqlConnection(this._connectionString))
-            {
-                SqlCommand SQLCmd = new SqlCommand();
-                SQLCmd.Connection = SQLConn;
-                SQLCmd.CommandText = StroredProcedureName;
-                SQLCmd.CommandType = CommandType.StoredProcedure;
-                //Loop for Paramets
-                for (int i = 0; i < ParaMeterCollection.Count; i++)
-                {
-                    SqlParameter sqlParaMeter = new SqlParameter();
-                    sqlParaMeter.IsNullable = true;
-                    sqlParaMeter.ParameterName = ParaMeterCollection[i].Key;
-                    sqlParaMeter.Value = ParaMeterCollection[i].Value;
-                    SQLCmd.Parameters.Add(sqlParaMeter);
-                }
-                //End of for loop                
-                SQLCmd = AddOutPutParametrofGivenType<T>(SQLCmd, OutPutParamerterName);
-                try
-                {
-                    await SQLConn.OpenAsync();
-                    await SQLCmd.ExecuteNonQueryAsync();
-                    return (T)SQLCmd.Parameters[OutPutParamerterName].Value;
-                }
-                catch (Exception e)
-                {
-                    throw e;
-                }
-                finally
-                {
-                    SQLConn.Close();
-                }
-            }
-        }
-
-        /// <summary>
-        /// Accept only int, Int16, long, DateTime, string (NVarcha of size  50),
-        /// bool, decimal ( of size 16,2), float
-        /// </summary>
-        /// <typeparam name="T">Return the given type of object</typeparam>
-        /// <param name="StroredProcedureName">Accet SQL procedure name in string</param>
-        /// <param name="ParaMeterCollection">Accept Key Value Collection for parameters</param>
-        /// <param name="OutPutParamerterName">Accept Output parameter for the stored procedures</param>
-        /// <returns></returns>
-        public async Task<T> ExecuteNonQueryAsGivenTypeAsync<T>(string StroredProcedureName, List<KeyValuePair<string, object>> ParaMeterCollection, string OutPutParamerterName, object OutPutParamerterValue)
-        {
-            using (SqlConnection SQLConn = new SqlConnection(this._connectionString))
-            {
-                SqlCommand SQLCmd = new SqlCommand();
-                SQLCmd.Connection = SQLConn;
-                SQLCmd.CommandText = StroredProcedureName;
-                SQLCmd.CommandType = CommandType.StoredProcedure;
-                //Loop for Paramets
-                for (int i = 0; i < ParaMeterCollection.Count; i++)
-                {
-                    SqlParameter sqlParaMeter = new SqlParameter();
-                    sqlParaMeter.IsNullable = true;
-                    sqlParaMeter.ParameterName = ParaMeterCollection[i].Key;
-                    sqlParaMeter.Value = ParaMeterCollection[i].Value;
-                    SQLCmd.Parameters.Add(sqlParaMeter);
-                }
-                //End of for loop                
-                SQLCmd = AddOutPutParametrofGivenType<T>(SQLCmd, OutPutParamerterName, OutPutParamerterValue);
-                try
-                {
-                    await SQLConn.OpenAsync();
-                    await SQLCmd.ExecuteNonQueryAsync();
-                    return (T)SQLCmd.Parameters[OutPutParamerterName].Value; ;
-                }
-                catch (Exception e)
-                {
-                    throw e;
-                }
-                finally
-                {
-                    SQLConn.Close();
-                }
-            }
-        }
 
         /// <summary>
         /// 
@@ -746,6 +393,11 @@ namespace SQLHelper
             if (typeof(T) == typeof(float))
             {
                 SQLCmd.Parameters.Add(new SqlParameter(OutPutParamerterName, SqlDbType.Float));
+                SQLCmd.Parameters[OutPutParamerterName].Direction = ParameterDirection.Output;
+            }
+            else if (typeof(T) == typeof(Guid))
+            {
+                SQLCmd.Parameters.Add(new SqlParameter(OutPutParamerterName, SqlDbType.UniqueIdentifier));
                 SQLCmd.Parameters[OutPutParamerterName].Direction = ParameterDirection.Output;
             }
             return SQLCmd;
@@ -803,227 +455,13 @@ namespace SQLHelper
                 SQLCmd.Parameters.Add(new SqlParameter(OutPutParamerterName, SqlDbType.Float));
                 SQLCmd.Parameters[OutPutParamerterName].Direction = ParameterDirection.Output;
             }
+            else if (typeof(T) == typeof(Guid))
+            {
+                SQLCmd.Parameters.Add(new SqlParameter(OutPutParamerterName, SqlDbType.UniqueIdentifier));
+                SQLCmd.Parameters[OutPutParamerterName].Direction = ParameterDirection.Output;
+            }
             SQLCmd.Parameters[OutPutParamerterName].Value = OutPutParamerterValue;
             return SQLCmd;
-        }
-
-        /// <summary>
-        /// Execute Non Query
-        /// </summary>
-        /// <param name="StroredProcedureName">Store Procedure Name In String</param>
-        /// <param name="ParaMeterCollection">Accept Key Value Collection For Parameters</param>
-        /// <param name="OutPutParamerterName">Accept OutPut Key Value Collection For Parameters</param>
-        /// <returns></returns>
-        public async Task<int> ExecuteNonQueryAsync(string StroredProcedureName, List<KeyValuePair<string, object>> ParaMeterCollection, string OutPutParamerterName)
-        {
-            using (SqlConnection SQLConn = new SqlConnection(this._connectionString))
-            {
-
-                using (SqlCommand SQLCmd = new SqlCommand())
-                {
-                    SQLCmd.Connection = SQLConn;
-                    SQLCmd.CommandText = StroredProcedureName;
-                    SQLCmd.CommandType = CommandType.StoredProcedure;
-                    //Loop for Paramets
-                    for (int i = 0; i < ParaMeterCollection.Count; i++)
-                    {
-                        SqlParameter sqlParaMeter = new SqlParameter();
-                        sqlParaMeter.IsNullable = true;
-                        sqlParaMeter.ParameterName = ParaMeterCollection[i].Key;
-                        sqlParaMeter.Value = ParaMeterCollection[i].Value;
-                        SQLCmd.Parameters.Add(sqlParaMeter);
-                    }
-                    //End of for loop
-                    SQLCmd.Parameters.Add(new SqlParameter(OutPutParamerterName, SqlDbType.Int));
-                    SQLCmd.Parameters[OutPutParamerterName].Direction = ParameterDirection.Output;
-                    try
-                    {
-                        await SQLConn.OpenAsync();
-                        await SQLCmd.ExecuteNonQueryAsync();
-                        int ReturnValue = (int)SQLCmd.Parameters[OutPutParamerterName].Value;
-                        SQLConn.Close();
-                        return ReturnValue;
-                    }
-                    catch (Exception e)
-                    {
-                        throw e;
-                    }
-                    finally
-                    {
-                        SQLConn.Close();
-                    }
-                }
-            }
-        }
-
-        /// <summary>
-        /// Execute Non Query
-        /// </summary>
-        /// <param name="StroredProcedureName"> Store Procedure Name</param>
-        /// <param name="ParaMeterCollection">Accept Key Value Collection For Parameters</param>
-        /// <param name="OutPutParamerterName">Accept Output Key Value Collection For Parameters</param>
-        /// <returns></returns>
-        public async Task<int> ExecuteNonQueryAsync(string StroredProcedureName, List<KeyValuePair<string, string>> ParaMeterCollection, string OutPutParamerterName)
-        {
-            using (SqlConnection SQLConn = new SqlConnection(this._connectionString))
-            {
-                using (SqlCommand SQLCmd = new SqlCommand())
-                {
-                    SQLCmd.Connection = SQLConn;
-                    SQLCmd.CommandText = StroredProcedureName;
-                    SQLCmd.CommandType = CommandType.StoredProcedure;
-                    //Loop for Paramets
-                    for (int i = 0; i < ParaMeterCollection.Count; i++)
-                    {
-                        SQLCmd.Parameters.Add(new SqlParameter(ParaMeterCollection[i].Key, ParaMeterCollection[i].Value));
-                    }
-                    //End of for loop
-                    SQLCmd.Parameters.Add(new SqlParameter(OutPutParamerterName, SqlDbType.Int));
-                    SQLCmd.Parameters[OutPutParamerterName].Direction = ParameterDirection.Output;
-                    try
-                    {
-                        await SQLConn.OpenAsync();
-                        await SQLCmd.ExecuteNonQueryAsync();
-                        int ReturnValue = (int)SQLCmd.Parameters[OutPutParamerterName].Value;
-                        SQLConn.Close();
-                        return ReturnValue;
-                    }
-                    catch (Exception e)
-                    {
-                        throw e;
-                    }
-                    finally
-                    {
-                        SQLConn.Close();
-                    }
-                }
-            }
-        }
-
-        /// <summary>
-        /// Execute Non Query
-        /// </summary>
-        /// <param name="StroredProcedureName">Store Procedure Name</param>
-        /// <param name="ParaMeterCollection">Accept Key Value Collection For Parameters</param>
-        /// <param name="OutPutParamerterName">Accept Output  For Parameters Name</param>
-        /// <param name="OutPutParamerterValue">Accept OutPut For Parameters Value</param>
-        /// <returns></returns>
-        public async Task<int> ExecuteNonQueryAsync(string StroredProcedureName, List<KeyValuePair<string, string>> ParaMeterCollection, string OutPutParamerterName, object OutPutParamerterValue)
-        {
-            using (SqlConnection SQLConn = new SqlConnection(this._connectionString))
-            {
-                using (SqlCommand SQLCmd = new SqlCommand())
-                {
-                    SQLCmd.Connection = SQLConn;
-                    SQLCmd.CommandText = StroredProcedureName;
-                    SQLCmd.CommandType = CommandType.StoredProcedure;
-                    //Loop for Paramets
-                    for (int i = 0; i < ParaMeterCollection.Count; i++)
-                    {
-                        SQLCmd.Parameters.Add(new SqlParameter(ParaMeterCollection[i].Key, ParaMeterCollection[i].Value));
-                    }
-                    //End of for loop
-                    SQLCmd.Parameters.Add(new SqlParameter(OutPutParamerterName, SqlDbType.Int));
-                    SQLCmd.Parameters[OutPutParamerterName].Direction = ParameterDirection.Output;
-                    SQLCmd.Parameters[OutPutParamerterName].Value = OutPutParamerterValue;
-                    try
-                    {
-                        await SQLConn.OpenAsync();
-                        await SQLCmd.ExecuteNonQueryAsync();
-                        int ReturnValue = (int)SQLCmd.Parameters[OutPutParamerterName].Value;
-                        SQLConn.Close();
-                        return ReturnValue;
-                    }
-                    catch (Exception e)
-                    {
-                        throw e;
-                    }
-                    finally
-                    {
-                        SQLConn.Close();
-                    }
-                }
-            }
-        }
-
-        /// <summary>
-        /// Execute Non Query
-        /// </summary>
-        /// <param name="StroredProcedureName">Store Procedure Name</param>
-        /// <param name="OutPutParamerterName">Accept Output For Parameters Name</param>
-        /// <returns></returns>
-        public async Task<int> ExecuteNonQueryAsync(string StroredProcedureName, string OutPutParamerterName)
-        {
-            using (SqlConnection SQLConn = new SqlConnection(this._connectionString))
-            {
-                using (SqlCommand SQLCmd = new SqlCommand())
-                {
-                    SQLCmd.Connection = SQLConn;
-                    SQLCmd.CommandText = StroredProcedureName;
-                    SQLCmd.CommandType = CommandType.StoredProcedure;
-                    SQLCmd.Parameters.Add(new SqlParameter(OutPutParamerterName, SqlDbType.Int));
-                    SQLCmd.Parameters[OutPutParamerterName].Direction = ParameterDirection.Output;
-
-                    try
-                    {
-                        await SQLConn.OpenAsync();
-                        await SQLCmd.ExecuteNonQueryAsync();
-                        int ReturnValue = (int)SQLCmd.Parameters[OutPutParamerterName].Value;
-                        SQLConn.Close();
-                        return ReturnValue;
-                    }
-                    catch (Exception e)
-                    {
-                        throw e;
-                    }
-                    finally
-                    {
-                        SQLConn.Close();
-                    }
-                }
-
-            }
-        }
-
-        /// <summary>
-        /// Execute Non Query
-        /// </summary>
-        /// <param name="StroredProcedureName">StoreProcedure Name</param>
-        /// <param name="OutPutParamerterName">Accept Output For Parameter Name</param>
-        /// <param name="OutPutParamerterValue">Accept Output For Parameter Value</param>
-        /// <returns></returns>
-        public async Task<int> ExecuteNonQueryAsync(string StroredProcedureName, string OutPutParamerterName, object OutPutParamerterValue)
-        {
-            using (SqlConnection SQLConn = new SqlConnection(this._connectionString))
-            {
-                using (SqlCommand SQLCmd = new SqlCommand())
-                {
-                    SQLCmd.Connection = SQLConn;
-                    SQLCmd.CommandText = StroredProcedureName;
-                    SQLCmd.CommandType = CommandType.StoredProcedure;
-                    SQLCmd.Parameters.Add(new SqlParameter(OutPutParamerterName, SqlDbType.Int));
-                    SQLCmd.Parameters[OutPutParamerterName].Direction = ParameterDirection.Output;
-                    SQLCmd.Parameters[OutPutParamerterName].Value = OutPutParamerterValue;
-
-                    try
-                    {
-                        await SQLConn.OpenAsync();
-                        await SQLCmd.ExecuteNonQueryAsync();
-                        int ReturnValue = (int)SQLCmd.Parameters[OutPutParamerterName].Value;
-                        SQLConn.Close();
-                        return ReturnValue;
-                    }
-                    catch (Exception e)
-                    {
-                        throw e;
-                    }
-                    finally
-                    {
-                        SQLConn.Close();
-                    }
-                }
-
-            }
         }
 
         /// <summary>
@@ -1052,51 +490,6 @@ namespace SQLHelper
                         sqlParaMeter.ParameterName = ParaMeterCollection[i].Key;
                         sqlParaMeter.Value = ParaMeterCollection[i].Value;
                         SQLCmd.Parameters.Add(sqlParaMeter);
-                    }
-                    //End of for loop
-
-                    SQLAdapter.SelectCommand = SQLCmd;
-                    try
-                    {
-                        await SQLConn.OpenAsync();
-                        SQLAdapter.Fill(SQLds);
-                        SQLConn.Close();
-                        return SQLds;
-                    }
-                    catch (Exception e)
-                    {
-                        throw e;
-                    }
-                    finally
-                    {
-                        SQLConn.Close();
-                    }
-                }
-            }
-        }
-
-        /// <summary>
-        /// Execute As DataSet
-        /// </summary>
-        /// <param name="StroredProcedureName">StoreProcedure Name</param>
-        /// <param name="ParaMeterCollection">Accept Key Value Collection For Parameters</param>
-        /// <returns></returns>
-        public async Task<DataSet> ExecuteAsDataSetAsync(string StroredProcedureName, List<KeyValuePair<string, string>> ParaMeterCollection)
-        {
-            using (SqlConnection SQLConn = new SqlConnection(this._connectionString))
-            {
-                using (SqlCommand SQLCmd = new SqlCommand())
-                {
-                    SqlDataAdapter SQLAdapter = new SqlDataAdapter();
-                    DataSet SQLds = new DataSet();
-                    SQLCmd.Connection = SQLConn;
-                    SQLCmd.CommandText = StroredProcedureName;
-                    SQLCmd.CommandType = CommandType.StoredProcedure;
-
-                    //Loop for Paramets
-                    for (int i = 0; i < ParaMeterCollection.Count; i++)
-                    {
-                        SQLCmd.Parameters.Add(new SqlParameter(ParaMeterCollection[i].Key, ParaMeterCollection[i].Value));
                     }
                     //End of for loop
 
@@ -1193,42 +586,6 @@ namespace SQLHelper
 
 
         /// <summary>
-        /// Execute As DataReader
-        /// </summary>
-        /// <param name="StroredProcedureName">StoreProcedure Name</param>
-        /// <param name="ParaMeterCollection">Accept Key Value Collection For Parameters</param>
-        /// <returns></returns>
-        public async Task<SqlDataReader> ExecuteAsDataReaderAsync(string StroredProcedureName, List<KeyValuePair<string, string>> ParaMeterCollection)
-        {
-            using (SqlConnection SQLConn = new SqlConnection(this._connectionString))
-            {                
-                SqlCommand SQLCmd = new SqlCommand();
-                SQLCmd.Connection = SQLConn;
-                SQLCmd.CommandText = StroredProcedureName;
-                SQLCmd.CommandType = CommandType.StoredProcedure;
-                //Loop for Paramets
-                for (int i = 0; i < ParaMeterCollection.Count; i++)
-                {
-                    SQLCmd.Parameters.Add(new SqlParameter(ParaMeterCollection[i].Key, ParaMeterCollection[i].Value));
-                }
-                //End of for loop
-                try
-                {
-                    SqlDataReader SQLReader;
-                    await SQLConn.OpenAsync();
-                    SQLReader = await SQLCmd.ExecuteReaderAsync(CommandBehavior.CloseConnection);
-                    return SQLReader;
-                }
-                catch (Exception e)
-                {
-                    throw e;
-                }
-            }
-
-        }
-
-
-        /// <summary>
         /// Execute As Object
         /// </summary>
         /// <typeparam name="T"><T></typeparam>
@@ -1252,60 +609,6 @@ namespace SQLHelper
                         sqlParaMeter.ParameterName = ParaMeterCollection[i].Key;
                         sqlParaMeter.Value = ParaMeterCollection[i].Value;
                         SQLCmd.Parameters.Add(sqlParaMeter);
-                    }
-                    //End of for loop
-                    try
-                    {
-                        SqlDataReader SQLReader;
-                        await SQLConn.OpenAsync();
-                        SQLReader = await SQLCmd.ExecuteReaderAsync(CommandBehavior.CloseConnection);
-                        ArrayList arrColl = DataSourceHelper.FillCollection(SQLReader, typeof(T));
-                        SQLConn.Close();
-                        if (SQLReader != null)
-                        {
-                            SQLReader.Close();
-                        }
-                        if (arrColl != null && arrColl.Count > 0)
-                        {
-                            return (T)arrColl[0];
-                        }
-                        else
-                        {
-                            return default(T);
-                        }
-                    }
-                    catch (Exception e)
-                    {
-                        throw e;
-                    }
-                    finally
-                    {
-                        SQLConn.Close();
-                    }
-                }
-            }
-        }
-
-        /// <summary>
-        /// Execute As Object
-        /// </summary>
-        /// <typeparam name="T"><T></typeparam>
-        /// <param name="StroredProcedureName">StoreProcedure Name</param>
-        /// <param name="ParaMeterCollection">Accept Key Value Collection For Parameters</param>
-        /// <returns></returns>
-        public async Task<T> ExecuteAsObjectAsync<T>(string StroredProcedureName, List<KeyValuePair<string, string>> ParaMeterCollection)
-        {
-            using (SqlConnection SQLConn = new SqlConnection(this._connectionString))
-            {
-                using (SqlCommand SQLCmd = new SqlCommand())
-                {
-                    SQLCmd.Connection = SQLConn;
-                    SQLCmd.CommandText = StroredProcedureName;
-                    SQLCmd.CommandType = CommandType.StoredProcedure;
-                    //Loop for Paramets
-                    for (int i = 0; i < ParaMeterCollection.Count; i++)
-                    {
-                        SQLCmd.Parameters.Add(new SqlParameter(ParaMeterCollection[i].Key, ParaMeterCollection[i].Value));
                     }
                     //End of for loop
                     try
@@ -1469,151 +772,6 @@ namespace SQLHelper
             }
         }
 
-        /// <summary>
-        /// Execute As list
-        /// </summary>
-        /// <typeparam name="T"></typeparam>
-        /// <param name="StroredProcedureName">StoreProcedure Name</param>
-        /// <param name="ParaMeterCollection"></param>
-        /// <returns></returns>
-        public async Task<List<T>> ExecuteAsListAsync<T>(string StroredProcedureName, List<KeyValuePair<string, object>> ParaMeterCollection)
-        {
-            using (SqlConnection SQLConn = new SqlConnection(this._connectionString))
-            {
-                using (SqlCommand SQLCmd = new SqlCommand())
-                {
-                    SQLCmd.Connection = SQLConn;
-                    SQLCmd.CommandText = StroredProcedureName;
-                    SQLCmd.CommandType = CommandType.StoredProcedure;
-                    //Loop for Paramets
-                    for (int i = 0; i < ParaMeterCollection.Count; i++)
-                    {
-                        SqlParameter sqlParaMeter = new SqlParameter();
-                        sqlParaMeter.IsNullable = true;
-                        sqlParaMeter.ParameterName = ParaMeterCollection[i].Key;
-                        sqlParaMeter.Value = ParaMeterCollection[i].Value;
-                        SQLCmd.Parameters.Add(sqlParaMeter);
-
-                    }
-                    //End of for loop
-                    try
-                    {
-                        SqlDataReader SQLReader;
-                        await SQLConn.OpenAsync();
-                        SQLReader = await SQLCmd.ExecuteReaderAsync(CommandBehavior.CloseConnection); //datareader automatically closes the SQL connection
-                        List<T> mList = new List<T>();
-                        mList = DataSourceHelper.FillCollection<T>(SQLReader);
-                        if (SQLReader != null)
-                        {
-                            SQLReader.Close();
-                        }
-                        SQLConn.Close();
-                        return mList;
-                    }
-                    catch (Exception e)
-                    {
-                        throw e;
-                    }
-                    finally
-                    {
-                        SQLConn.Close();
-                    }
-                }
-            }
-        }
-
-        /// <summary>
-        /// Execute As List
-        /// </summary>
-        /// <typeparam name="T"><T></typeparam>
-        /// <param name="StroredProcedureName">Store Procedure Name</param>
-        /// <param name="ParaMeterCollection">Accept Key Value Collection For Parameters</param>
-        /// <returns></returns>
-        public async Task<List<T>> ExecuteAsListAsync<T>(string StroredProcedureName, List<KeyValuePair<string, string>> ParaMeterCollection)
-        {
-            using (SqlConnection SQLConn = new SqlConnection(this._connectionString))
-            {
-                using (SqlCommand SQLCmd = new SqlCommand())
-                {
-                    SQLCmd.Connection = SQLConn;
-                    SQLCmd.CommandText = StroredProcedureName;
-                    SQLCmd.CommandType = CommandType.StoredProcedure;
-                    //Loop for Paramets
-                    for (int i = 0; i < ParaMeterCollection.Count; i++)
-                    {
-                        SqlParameter sqlParaMeter = new SqlParameter();
-                        sqlParaMeter.IsNullable = true;
-                        sqlParaMeter.ParameterName = ParaMeterCollection[i].Key;
-                        sqlParaMeter.Value = ParaMeterCollection[i].Value;
-                        SQLCmd.Parameters.Add(sqlParaMeter);
-                    }
-                    //End of for loop
-                    try
-                    {
-                        SqlDataReader SQLReader;
-                        await SQLConn.OpenAsync();
-                        SQLReader = await SQLCmd.ExecuteReaderAsync(CommandBehavior.CloseConnection); //datareader automatically closes the SQL connection
-                        List<T> mList = new List<T>();
-                        mList = DataSourceHelper.FillCollection<T>(SQLReader);
-                        if (SQLReader != null)
-                        {
-                            SQLReader.Close();
-                        }
-                        SQLConn.Close();
-                        return mList;
-                    }
-                    catch (Exception e)
-                    {
-                        throw e;
-                    }
-                    finally
-                    {
-                        SQLConn.Close();
-                    }
-                }
-            }
-        }
-
-        /// <summary>
-        /// Execute As List
-        /// </summary>
-        /// <typeparam name="T"><T></typeparam>
-        /// <param name="StroredProcedureName">Storedprocedure Name</param>
-        /// <returns></returns>
-        public async Task<List<T>> ExecuteAsListAsync<T>(string StroredProcedureName)
-        {
-            using (SqlConnection SQLConn = new SqlConnection(this._connectionString))
-            {
-                using (SqlCommand SQLCmd = new SqlCommand())
-                {
-                    SQLCmd.Connection = SQLConn;
-                    SQLCmd.CommandText = StroredProcedureName;
-                    SQLCmd.CommandType = CommandType.StoredProcedure;
-                    try
-                    {
-                        SqlDataReader SQLReader;
-                        await SQLConn.OpenAsync();
-                        SQLReader = await SQLCmd.ExecuteReaderAsync(CommandBehavior.CloseConnection); //datareader automatically closes the SQL connection
-                        List<T> mList = new List<T>();
-                        mList = DataSourceHelper.FillCollection<T>(SQLReader);
-                        if (SQLReader != null)
-                        {
-                            SQLReader.Close();
-                        }
-                        SQLConn.Close();
-                        return mList;
-                    }
-                    catch (Exception e)
-                    {
-                        throw e;
-                    }
-                    finally
-                    {
-                        SQLConn.Close();
-                    }
-                }
-            }
-        }
 
         /// <summary>
         /// Execute As Scalar 
@@ -1721,32 +879,6 @@ namespace SQLHelper
                         paramCollection.Add(new KeyValuePair<string, object>("@" + objProperty.Name.ToString(), objProperty.GetValue(obj, null)));
                     }
                     return paramCollection;
-                }
-            }
-            catch (Exception e)
-            {
-                throw e;
-            }
-            return paramCollection;
-        }
-
-        /// <summary>
-        /// Bulid Collection of List<KeyValuePair<string, string>> for Given object
-        /// </summary>
-        /// <typeparam name="List">List of Type(string,string)</typeparam>
-        /// <param name="paramCollection">List of Type(string,string)</param>
-        /// <param name="obj">Object</param>        
-        /// <returns> Collection of KeyValuePair<string, string> </returns>
-        public List<KeyValuePair<string, string>> BuildParameterCollection(List<KeyValuePair<string, string>> paramCollection, object obj)
-        {
-            try
-            {
-                foreach (PropertyInfo objProperty in obj.GetType().GetProperties())
-                {
-                    if (objProperty.GetValue(obj, null).ToString() != null)
-                    {
-                        paramCollection.Add(new KeyValuePair<string, string>("@" + objProperty.Name.ToString(), objProperty.GetValue(obj, null).ToString()));
-                    }
                 }
             }
             catch (Exception e)
